+++ src/main/java/org/jabref/gui/BasePanel.java
@@ -749,6 +749,19 @@ public class BasePanel extends JPanel implements ClipboardOwner {
         actions.put(Actions.REMOVE_FROM_GROUP, new GroupAddRemoveDialog(this, false, false));
         actions.put(Actions.MOVE_TO_GROUP, new GroupAddRemoveDialog(this, true, true));
 
+        actions.put(Actions.INTERSECTION_PREF_GROUP, (BaseAction) () -> {
+            boolean intersectionToggled = Globals.prefs.getBoolean(JabRefPreferences.GROUP_INTERSECT_SELECTIONS);
+            boolean newState = !intersectionToggled;
+            Globals.prefs.putBoolean(JabRefPreferences.GROUP_INTERSECT_SELECTIONS, newState);
+
+            if (newState) {
+                output(Localization.lang("Group view mode set to intersection"));
+            } else {
+                output(Localization.lang("Group view mode set to union"));
+            }
+
+        });
+
         actions.put(Actions.DOWNLOAD_FULL_TEXT, new FindFullTextAction(this));
     }
 
@@ -1086,7 +1099,7 @@ public class BasePanel extends JPanel implements ClipboardOwner {
     }
 
     private boolean saveDatabase(File file, boolean selectedOnly, Charset enc,
-                                 SavePreferences.DatabaseSaveType saveType)
+            SavePreferences.DatabaseSaveType saveType)
             throws SaveException {
         SaveSession session;
         frame.block();